mockito + some unit test

+++ src/main/java/controller/ConfigManager.java
@@ -24,7 +24,8 @@ public class ConfigManager {
                 PGConfigDelta configDelta = new PGConfigDelta(name, setting, unit, varType);
                 if (varType.equals("integer") || varType.equals("real")) {
                     int minValue = rs.getInt("min_val");
-                    long maxValue = Double.valueOf(rs.getString("max_val")).longValue();
+                    //String test = rs.getString("max_val");
+                    long maxValue = Long.valueOf(rs.getString("max_val")).longValue();
                     configDelta.setRange(minValue, maxValue);
                 }
                 if (varType.equals("enum")) {
